code_optimization/requirements.txt
absl-py==2.1.0
chex==0.1.86
etils==1.8.0
flax==0.8.3
fsspec==2024.3.1
importlib_resources==6.4.0
jax==0.4.28
jax-cuda12-pjrt==0.4.28
jax-cuda12-plugin==0.4.28
jaxlib==0.4.28
markdown-it-py==3.0.0
mdurl==0.1.2
ml-dtypes==0.4.0
msgpack==1.0.8
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.4.131
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.2.1.3
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.11
protobuf==5.26.1
Pygments==2.18.0
PyYAML==6.0.1
rich==13.7.1
setuptools==69.5.1
tensorstore==0.1.59
toolz==0.12.1
typing_extensions==4.11.0
zipp==3.18.1
